<==== Mercurial data ====>
Branch: master

+++ mediaserver/src/rest/handlers/business_event_log_rest_handler.cpp
@@ -1,7 +1,5 @@
 #include "business_event_log_rest_handler.h"
 
-#include "api/serializer/pb_serializer.h"
-
 #include "business/actions/abstract_business_action.h"
 
 #include <core/resource/camera_resource.h>
@@ -61,7 +59,7 @@ int QnBusinessEventLogRestHandler::executeGet(const QString& path, const QnReque
                     errStr = QString("Invalid action type %1. Valid range is [0..%2]").arg(params[i].second).arg(BusinessActionType::NotDefined-1);
             }
             else if (params[i].first == "brule_id") {
-                businessRuleId = params[i].second.toInt();
+                businessRuleId = params[i].second;
             }
         }
     }
